X/ptolstov/4 1 rdep fix build time artifacts synchronization

+++ build_utils/python/rdep_cmake.py
@@ -4,32 +4,45 @@ import sys
 import posixpath
 import argparse
 import glob
+from pathlib import Path
+
 from rdep import Rdep
 
 class RdepSyncher:
-    def __init__(self, rdep_repo_root):
+    def __init__(self, rdep_repo_root, verbose=False, sync_timestamps=False):
         self.versions = {}
         self.locations = {}
         self.rdep_target = None
         self.use_local = False
+        self.have_updated_packages = False
 
         self.rdep = Rdep(rdep_repo_root)
+        self.rdep.verbose = verbose
         self.rdep.fast_check = True
-        self.rdep.load_timestamps_for_fast_check()
 
         self._exported_paths = {}
+        self._env_exported_paths = {}
         self._synched_package_dirs = []
         self._include_ignored_dirs = []
 
-    def sync(self, package, path_variable=None, optional=False, do_not_include=False):
+        if sync_timestamps:
+            self.rdep.sync_timestamps()
+
+        self.rdep.load_timestamps_for_fast_check()
+
+    def sync(self,
+            package,
+            path_variable=None,
+            env_path_variable=None,
+            optional=False,
+            do_not_include=False):
         target, pack = posixpath.split(package)
 
         self.rdep.targets = [target] if target else [self.rdep_target]
 
         version = self.versions.get(pack)
-        path = self.locations.get(pack)
-
         full_package_name = pack + "-" + version if version else pack
+        path = self.locations.get(pack)
 
         if not path:
             sync_func = self.rdep.locate_package if self.use_local else self.rdep.sync_package
@@ -48,12 +61,19 @@ class RdepSyncher:
         self._synched_package_dirs.append(path)
         if path_variable:
             self._exported_paths[path_variable] = path
+        if env_path_variable:
+            self._env_exported_paths[env_path_variable] = path
         if do_not_include:
             self._include_ignored_dirs.append(path)
+
+        self.have_updated_packages = self.rdep.is_package_updated(target, full_package_name)
 
         return True
 
     def generate_cmake_include(self, file_name):
+        if Path(file_name).exists() and not self.have_updated_packages:
+            return
+
         with open(file_name, "w") as f:
             f.write("# Package versions.\n")
             for k, v in self.versions.items():
@@ -62,6 +82,10 @@ class RdepSyncher:
             f.write("\n# Exported package directories.\n")
             for k, v in self._exported_paths.items():
                 f.write("set({0} \"{1}\")\n".format(k, v))
+
+            f.write("\n# Package directories exported to environment.\n")
+            for k, v in self._env_exported_paths.items():
+                f.write("set(ENV{{{0}}} \"{1}\")\n".format(k, v))
 
             f.write("\n# List of synchronized package directories.\n")
             f.write("set(synched_package_dirs\n")
